backend/requirements.txt
fastapi
uvicorn
aiogram
supabase
python-dotenv
google-generativeai
google-api-python-client
youtube-transcript-api
pydantic
requests
